WallJumping implemented
Features:
Enum that controls the special movement of the character. That way, the character cannot be walking, climbing, walljumping simultaneously.
WallJumping: Character can wallslide if their y-velocity is less than zero and they touch a wall. They can jump by being in a wall-slide state and pressing jump or they can jump to a wall and immediately wall jump off it.
If the player stops holding a direction or they hold a direction that is normal to the wall they are touching, they will break free from the wall-slide state. This can be a bit buggy sometimes.