Climbing
There are two different types of climbing:
Climbing that allows horizontal movement, and climbing that has a top. If the Surface has a top, then the player will stop when they reach the top of the surface., otherwise, the game will expect the player to collide with a ceiling or platform.

Future Stuff: Jumping while on climbable surfaces.